Implemented review feedback
