+++ processor/standalone/pom.xml
@@ -32,7 +32,7 @@
         <dependency>
             <groupId>org.opennms.oce</groupId>
             <artifactId>processor-api</artifactId>
-            <version>1.0.0-SNAPSHOT</version>
+            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
